apply max img size to SVG: https://github.com/metanorma/metanorma-iso/issues/814
